Time: 0 ms (100.00%), Space: 40.7 MB (6.21%) - LeetHub

+++ 2582-pass-the-pillow/2582-pass-the-pillow.java
@@ -1,7 +1,7 @@
 class Solution {
     public int passThePillow(int n, int time) {
-        int i = 1;
-        int s = 1;
+        short i = 1;
+        short s = 1;
         while ( time > 0 ) {
             int distance = Math.min(time,n-1);
             i += distance * s;
